src/test/kotlin/nl/rjcoding/aoc2020/Day24Tests.kt
package nl.rjcoding.aoc2020

import kotlin.test.Test
import kotlin.test.assertEquals

class Day24Tests {

    val input = """
        sesenwnenenewseeswwswswwnenewsewsw
        neeenesenwnwwswnenewnwwsewnenwseswesw
        seswneswswsenwwnwse
        nwnwneseeswswnenewneswwnewseswneseene
        swweswneswnenwsewnwneneseenw
        eesenwseswswnenwswnwnwsewwnwsene
        sewnenenenesenwsewnenwwwse
        wenwwweseeeweswwwnwwe
        wsweesenenewnwwnwsenewsenwwsesesenwne
        neeswseenwwswnwswswnw
        nenwswwsewswnenenewsenwsenwnesesenew
        enewnwewneswsewnwswenweswnenwsenwsw
        sweneswneswneneenwnewenewwneswswnese
        swwesenesewenwneswnwwneseswwne
        enesenwswwswneneswsenwnewswseenwsese
        wnwnesenesenenwwnenwsewesewsesesew
        nenewswnwewswnenesenwnesewesw
        eneswnwswnwsenenwnwnwwseeswneewsenese
        neswnwewnwnwseenwseesewsenwsweewe
        wseweeenwnesenwwwswnew
    """.trimIndent().lineSequence()

    @Test
    fun flip() {
        val instructions = Day24.parse(input)
        val result = Day24.flip(instructions)
        assertEquals(10, result.size)
    }

    @Test
    fun next() {
        val instructions = Day24.parse(input)
        val init = Day24.flip(instructions)
        val day100 = Day24.iterate(init, 100)
        assertEquals(2208, day100.size)
    }
}